styled nav bar
