+++ src/index.html
@@ -4,14 +4,13 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="shortcut icon" href="#">
-  <title>Test</title>
+  <title>Gabrielle Tobermann</title>
 </head>
 <body>
   <body>
     <!-- Navigation -->
     <nav>
       <!-- TODO: Insert some sort of H1-H6 with your name here -->
-      <h1>Gabrielle Tobermann</h1>
       <ul id="navLinks">
         <li><a id="navToBio">Bio</a></li>
         <li><a id="navToTechnologies">Technologies</a></li>
@@ -19,6 +18,10 @@
       </ul>
     </nav>
     <!-- End Navigation -->
+    <header class="header">
+      <h1 class="name-header">Gabrielle Tobermann </h1>
+      <h2 class="name-header">Web Developer</h2>
+    </header>
 
     <!-- Primary View -->
     <div class="fullPage" id="bioPage">
