--- src/javascripts/helpers/signIn.js
@@ -1,9 +1,0 @@
-import firebase from 'firebase/app';
-import 'firebase/auth';
-
-const signIn = () => {
-  const provider = new firebase.auth.GoogleAuthProvider();
-  firebase.auth().signInWithPopup(provider);
-};
-
-export default signIn;
