+++ src/styles/main.scss
@@ -1,19 +1,15 @@
 @import "~bootstrap/scss/bootstrap";
 @import "~@fortawesome/fontawesome-free/css/all.min.css";
+@import url('https://fonts.googleapis.com/css2?family=Satisfy&display=swap');
 
 body {
   background-color: white;
 }
 
 nav {
-  background-color: #004221;
-  color: white;
-  
-}
-
-h1 {
-  padding: 30px;
-  font-family: 'Times New Roman', Times, serif;
+  height: 120px;
+  background-image: linear-gradient(180deg, #ffffff 36.67%, #f0f0f0 36.67%, #f0f0f0 50%, #ffffff 50%, #ffffff 86.67%, #f0f0f0 86.67%, #f0f0f0 100%);
+  background-size: 30.00px 30.00px; 
 }
 
 #navLinks {
@@ -24,11 +20,20 @@ h1 {
 }
 
 #navLinks li {
+  padding-top: 25px;
   padding-left:50px;
   padding-right: 50px;
-  padding-bottom: 20px;
-  font-size: 20px;
-  font-family: 'Courgette', cursive;
+  font-size: 40px;
+  font-family: 'Satisfy', cursive;
+}
+
+header {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-family: 'Times New Roman', Times, serif;
+  height: 400px;
 }
 
 #bioPage {
